Fix error on unsupported Lua type function
This patch fixes an error in the console when Tarantool is unable to
encode function.

Reproduce the bug as follows.
```
> box.session
unsupported Lua type 'function' (LuajitError, code 0x20)...
```

It turns out Tarantool tries to encode some of the fields as functions.
It can be fixed by providing a configuration to `yaml` Lua module on the
server side.

```lua
yaml.cfg{ encode_use_tostring = true }
```

This is just a temporary fix which uses the same hack as tt.

Reported by vladgrubov@gmail.com (@ochaton)